# backend/openapi.yaml
openapi: 3.1.0
info:
    title: VidyaSetu
    version: 0.1.0
    description: |
        VidyaSetu is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        Page_AssessmentResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/AssessmentResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[AssessmentResponse]
            type: object
...
        Page_EnrollmentResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/EnrollmentResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[EnrollmentResponse]
            type: object
...
        Page_SessionResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/SessionResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[SessionResponse]
            type: object
...
        Page_WorkshopResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/WorkshopResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[WorkshopResponse]
            type: object
...
        WorkshopAnalyticsResponse:
            properties:
                assessment:
                    $ref: "#/components/schemas/WorkshopAssessmentAnalytics"
                    default:
                        avg_percentage: 0.0
                        avg_score: 0.0
                        pass_rate_percentage: 0.0
                        total_submissions: 0
                attendance:
                    $ref: "#/components/schemas/WorkshopAttendanceAnalytics"
                    default:
                        avg_attendance_percentage: 0.0
                        total_attendance_records: 0
                enrollment:
                    $ref: "#/components/schemas/WorkshopEnrollmentAnalytics"
                    default:
                        active: 0
                        completed: 0
                        dropped: 0
                        total_enrolled: 0
                workshop_id:
                    title: Workshop Id
                    type: string
            required:
                - workshop_id
            title: WorkshopAnalyticsResponse
            type: object
...
        WorkshopAttendanceAnalytics:
            properties:
                avg_attendance_percentage:
                    default: 0.0
                    title: Avg Attendance Percentage
                    type: number
                total_attendance_records:
                    default: 0
                    title: Total Attendance Records
                    type: integer
            title: WorkshopAttendanceAnalytics
            type: object
        WorkshopCreate:
            properties:
                description:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Description
                end_date:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: End Date
                institution_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Institution Id
                start_date:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Start Date
                title:
                    title: Title
                    type: string
            required:
                - title
            title: WorkshopCreate
            type: object
...
        WorkshopLeaderboardResponse:
            properties:
                entries:
                    default: []
                    items:
                        $ref: "#/components/schemas/LeaderboardEntry"
                    title: Entries
                    type: array
                workshop_id:
                    title: Workshop Id
                    type: string
            required:
                - workshop_id
            title: WorkshopLeaderboardResponse
            type: object
...
        WorkshopUpdate:
            properties:
                description:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Description
                end_date:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: End Date
                start_date:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Start Date
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
            title: WorkshopUpdate
            type: object
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/analytics/leaderboard/workshop/{workshop_id}:
        get:
            operationId: workshop_leaderboard_api_v1_analytics_leaderboard_workshop__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopLeaderboardResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Workshop leaderboard ranked by average percentage across workshop assessments
            tags:
                - analytics
...
    /api/v1/analytics/workshop/{workshop_id}:
        get:
            operationId: workshop_analytics_api_v1_analytics_workshop__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopAnalyticsResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get cohort-level averages and pass rates for a workshop (staff only)
            tags:
                - analytics
...
    /api/v1/assessments/workshop/{workshop_id}:
        get:
            operationId: list_by_workshop_api_v1_assessments_workshop__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_AssessmentResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List assessments for a workshop
            tags:
                - assessments
...
    /api/v1/enrollments/workshop/{workshop_id}:
        get:
            operationId: list_by_workshop_api_v1_enrollments_workshop__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_EnrollmentResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get all enrollments for a workshop (staff only)
            tags:
                - enrollments
...
    /api/v1/sessions/workshop/{workshop_id}:
        get:
            operationId: list_by_workshop_api_v1_sessions_workshop__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_SessionResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List sessions for a workshop
            tags:
                - sessions
...
    /api/v1/workshops/:
        get:
            operationId: list_workshops_api_v1_workshops__get
            parameters:
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_WorkshopResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List workshops (scoped by institution for non-admins)
            tags:
                - workshops
        post:
            operationId: create_api_v1_workshops__post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkshopCreate"
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Create a workshop (admin / institution_admin / educator)
            tags:
                - workshops
    /api/v1/workshops/{workshop_id}:
        delete:
            operationId: delete_one_api_v1_workshops__workshop_id__delete
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            responses:
                "204":
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Delete a workshop (admin / institution_admin only)
            tags:
                - workshops
        get:
            operationId: get_one_api_v1_workshops__workshop_id__get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get a workshop by ID
            tags:
                - workshops
        patch:
            operationId: update_one_api_v1_workshops__workshop_id__patch
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkshopUpdate"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Update a workshop
            tags:
                - workshops
    /api/v1/workshops/{workshop_id}/educator-profile:
        get:
            operationId: get_workshop_educator_profile_api_v1_workshops__workshop_id__educator_profile_get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkshopEducatorProfileResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get primary educator profile for a workshop
            tags:
                - workshops
    /api/v1/workshops/{workshop_id}/modules:
        get:
            operationId: list_modules_api_v1_workshops__workshop_id__modules_get
            parameters:
                - in: path
                  name: workshop_id
                  required: true
                  schema:
                      title: Workshop Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_ModuleResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List modules for a workshop
            tags:
                - workshops